reactor(node-sdk): Refactor node SDK to Typescript
Closes issue #25
