+++ .github/workflows/ci.yml
@@ -104,6 +104,7 @@ jobs:
         run: |
           cd node
           npm i
+          npm run build
           LD_LIBRARY_PATH=/usr/local/lib node example.js
 
       - name: Test Rust Host SDK
