+++ node/package.json
@@ -12,7 +12,8 @@
   "author": "The Extism Authors <oss@extism.org>",
   "license": "BSD-3-Clause",
   "private": false,
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "type": "module",
   "homepage": "https://extism.org",
   "repository": {
@@ -20,12 +21,19 @@
     "url": "https://github.com/extism/extism.git"
   },
   "scripts": {
-    "example": "node example.js"
+    "prepare" : "npm run build",
+    "example": "node example.js",
+    "build": "tsc"
   },
   "dependencies": {
     "ffi-napi": "^4.0.3"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "devDependencies": {
+    "@types/ffi-napi": "^4.0.6",
+    "@types/node": "^18.11.4",
+    "typescript": "^4.8.4"
   }
 }
